# [P4_FUZZER] Check resource limits in fuzzer. For now we only insert, to get to full tables faster. Running 10k iterations now.